Store ticket period as unsigned long. The usage of the library was limited due to this

+++ src/Ticker.cpp
@@ -24,7 +24,7 @@
 
 #include "Ticker.h"
 
-jsc::Ticker::Ticker(uint16_t tickerTime): tickerTime(tickerTime) {
+jsc::Ticker::Ticker(unsigned long tickerTime): tickerTime(tickerTime) {
 }
 
 uint16_t jsc::Ticker::perform() {
